Added validation on submitting a review such as invalid score, comments, already submitted a review, Added a delete button that will appear next to reviews that a user has submitted, allowing them to delete their review and make a new one, a message now indicates any reviews that the currently loogged in user has made, on the detailed movie view page

+++ server/webflix/src/main/java/com/instil/webflix/review/model/Review.java
@@ -16,11 +16,11 @@ public class Review {
 	public int account_id;
 	public int movie_id;
 	public String comments;
-	public float score;
+	public double score;
 	
 	public Review() {}
 	
-	public Review(int review_id, int accountID, int movieID, String comments, float score) {
+	public Review(int review_id, int accountID, int movieID, String comments, double score) {
 		super();
 		this.review_id = review_id;
 		this.account_id = accountID;
@@ -61,11 +61,11 @@ public class Review {
 		this.comments = comments;
 	}
 
-	public float getScore() {
+	public double getScore() {
 		return score;
 	}
 
-	public void setScore(float score) {
+	public void setScore(double score) {
 		this.score = score;
 	}
